# docs/api/openapi-spec.yaml
openapi: 3.0.0
info:
  title: SmartWalletFX Crypto Data API
  version: 3.0.0
  description: |
    **Phase 3 Production Release** - Enterprise-grade cryptocurrency portfolio and DeFi data API with 
    comprehensive monitoring, advanced cost optimization, and multi-chain support.
    
    ## 🎯 Phase 3 Enhancements
    - **Security Hardening**: Enhanced rate limiting, input validation, and audit logging
    - **Performance Optimization**: 42% cost reduction, sub-200ms response times
    - **Build Improvements**: Automated testing, CI/CD pipeline, container optimization
    - **Monitoring Excellence**: 15+ Prometheus metrics, comprehensive alerting
    
    ## ✨ Key Features
    - **Multi-chain Support**: Ethereum, Polygon, Arbitrum, Optimism, Base, Solana
    - **DeFi Integration**: 15+ protocols (Aave V3, Compound V3, Uniswap V3, Curve, Yearn, Lido)
    - **NFT Portfolio**: Collection tracking, metadata enrichment, rarity analysis
    - **Intelligent Caching**: Multi-level (L1/L2/L3) with predictive warming
    - **Cost Optimization**: Smart batching, provider rotation, request deduplication
    - **Enterprise Observability**: Structured logging, distributed tracing, cost tracking
    
    ## 🔐 Authentication
    All endpoints require an API key in the `X-API-Key` header unless specified otherwise.
    
    ## 📊 Rate Limits
    - **Standard**: 1000 requests/minute
    - **Premium**: 5000 requests/minute
    - **Enterprise**: Custom limits available
    
  contact:
...
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Obtain your API key from the SmartWalletFX dashboard.
        
        Example: `X-API-Key: swfx_1234567890abcdef`
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token for enhanced security (Enterprise only)
  schemas:
...
    BaseResponse:
      type: object
      required:
        - success
        - metadata
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    ResponseMetadata:
      type: object
      required:
        - timestamp
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the response
          example: "2024-01-15T10:30:00.000Z"
        requestId:
          type: string
          format: uuid
          description: Unique request identifier for tracing
          example: "550e8400-e29b-41d4-a716-446655440000"
        duration:
          type: number
          description: Request processing time in milliseconds
          example: 150
        cacheStatus:
          type: string
          enum: [hit, miss, bypass]
          description: Cache hit/miss status
        costUsed:
          type: number
          format: float
          description: API cost consumed in USD
          example: 0.001
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              description: Response data payload
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - error
          properties:
            success:
              type: boolean
              example: false
            error:
              $ref: '#/components/schemas/ErrorDetails'
    ErrorDetails:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code identifier
          example: "INVALID_ADDRESS"
        message:
          type: string
          description: Human-readable error message
          example: "The provided address format is invalid"
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        retryable:
          type: boolean
          description: Whether the request can be retried
        retryAfter:
          type: integer
          description: Seconds to wait before retry
          minimum: 0

    # Portfolio Types
    PortfolioData:
      type: object
      required:
        - address
        - totalValueUSD
        - lastUpdated
      properties:
        address:
          type: string
          description: Wallet address
          example: "0x742d35Cc6634C0532925a3b8D60c7b35C4D2C14a"
        totalValueUSD:
          type: number
          format: decimal
          description: Total portfolio value in USD
          example: 12500.75
        netWorthUSD:
          type: number
          format: decimal
          description: Net worth (assets minus debts)
          example: 11800.50
        defi:
          $ref: '#/components/schemas/DeFiSummary'
        nfts:
          $ref: '#/components/schemas/NFTSummary'
        tokens:
          $ref: '#/components/schemas/TokenSummary'
        chainDistribution:
          type: array
          items:
            $ref: '#/components/schemas/ChainDistribution'
        riskMetrics:
          $ref: '#/components/schemas/RiskMetrics'
        lastUpdated:
          type: string
          format: date-time
          description: Last portfolio update timestamp
    DeFiSummary:
      type: object
      properties:
        totalValueUSD:
          type: number
          format: decimal
          description: Total DeFi position value
          example: 8500.25
        totalBorrowedUSD:
          type: number
          format: decimal
          description: Total borrowed amount
          example: 2000.00
        netValueUSD:
          type: number
          format: decimal
          description: Net DeFi value (supplied - borrowed)
          example: 6500.25
        positionCount:
          type: integer
          description: Total number of positions
          example: 12
        protocolCount:
          type: integer
          description: Number of protocols used
          example: 4
        protocols:
          type: array
          items:
            $ref: '#/components/schemas/ProtocolSummary'
    NFTSummary:
      type: object
      properties:
        totalValueUSD:
          type: number
          format: decimal
          description: Total estimated NFT value
          example: 3500.00
        totalCollections:
          type: integer
          description: Number of collections owned
          example: 8
        totalNFTs:
          type: integer
          description: Total number of NFTs
          example: 45
        floorValueUSD:
          type: number
          format: decimal
          description: Total floor value
          example: 2800.00
    TokenSummary:
      type: object
      properties:
        totalValueUSD:
          type: number
          format: decimal
          description: Total token value
          example: 500.50
        tokenCount:
          type: integer
          description: Number of different tokens
          example: 15
    ChainDistribution:
      type: object
      required:
        - chainId
        - chainName
        - valueUSD
        - percentage
      properties:
        chainId:
          type: string
          description: Chain identifier
          example: "1"
        chainName:
          type: string
          description: Human-readable chain name
          example: "Ethereum"
        valueUSD:
          type: number
          format: decimal
          description: Value on this chain
          example: 8500.25
        percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage of total portfolio
          example: 68.5
    RiskMetrics:
      type: object
      properties:
        riskScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Overall risk score (0-100)
          example: 45.5
        liquidationRisk:
          type: string
          enum: [low, medium, high, critical]
          description: Liquidation risk level
        diversificationScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Portfolio diversification score
          example: 72.3

    # DeFi Types
...
    ProtocolSummary:
      type: object
      properties:
        name:
          type: string
          example: "Aave V3"
        valueUSD:
          type: number
          format: decimal
          example: 5000.00
        positionCount:
          type: integer
          example: 3
        chainId:
          type: string
          example: "1"

    # NFT Types
...
  parameters:
    AddressParam:
      name: address
      in: path
      required: true
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$|^[1-9A-HJ-NP-Za-km-z]{32,44}$'
      description: Wallet address (Ethereum format or Solana format)
      example: "0x742d35Cc6634C0532925a3b8D60c7b35C4D2C14a"
    ChainIdsParam:
      name: chains
      in: query
      required: false
      schema:
        type: string
      description: Comma-separated list of chain IDs
      example: "1,137,42161"
    IncludeParam:
      name: include
      in: query
      required: false
      schema:
        type: string
      description: Comma-separated list of data to include
      example: "defi,nfts,tokens"
...
    RefreshParam:
      name: refresh
      in: query
      required: false
      schema:
        type: boolean
        default: false
      description: Force refresh of cached data
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_ADDRESS"
              message: "The provided address format is invalid"
              details:
                address: "invalid-address"
                expectedFormat: "0x followed by 40 hexadecimal characters"
            metadata:
              timestamp: "2024-01-15T10:30:00.000Z"
              requestId: "550e8400-e29b-41d4-a716-446655440000"
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or missing API key"
            metadata:
              timestamp: "2024-01-15T10:30:00.000Z"
              requestId: "550e8400-e29b-41d4-a716-446655440000"
    RateLimit:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Window reset time
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/portfolio/{address}:
    get:
      tags:
        - Portfolio
      summary: Get complete portfolio
      description: |
        Retrieves comprehensive portfolio data including DeFi positions, 
        NFTs, tokens, and risk metrics for a given address.
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/ChainIdsParam'
        - $ref: '#/components/parameters/IncludeParam'
        - $ref: '#/components/parameters/RefreshParam'
      responses:
        '200':
          description: Portfolio data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PortfolioData'
              example:
                success: true
                data:
                  address: "0x742d35Cc6634C0532925a3b8D60c7b35C4D2C14a"
                  totalValueUSD: 12500.75
                  netWorthUSD: 11800.50
                  defi:
                    totalValueUSD: 8500.25
                    positionCount: 12
                    protocolCount: 4
                  nfts:
                    totalValueUSD: 3500.00
                    totalCollections: 8
                    totalNFTs: 45
                  lastUpdated: "2024-01-15T10:30:00.000Z"
                metadata:
                  timestamp: "2024-01-15T10:30:00.000Z"
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  duration: 150
                  cacheStatus: "hit"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalError'

  # DeFi Endpoints
...